refactor icons
