--- frontend/src/components/ImgIcon/ImgIcon.scss
@@ -1,23 +1,0 @@
-.img {
-  user-select: none;
-  border-radius: 50%;
-
-  &.sm {
-    width: 1rem;
-    height: 1rem;
-  }
-
-  &.md {
-    width: 2rem;
-    height: 2rem;
-  }
-
-  &.lg {
-    width: 4rem;
-    height: 4rem;
-  }
-
-  &.bg-white {
-    background: white;
-  }
-}
